# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Dash Server API",
        "description": "Api de l'application Dash",
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/get_month_invoices": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "description": "Récupérer les factures du mois",
                "operationId": "fd02603120789e695269d40a33d52ad0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "month",
                                    "year"
                                ],
                                "properties": {
                                    "month": {
                                        "type": "integer"
                                    },
                                    "year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateurs récupérés avec succès"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/v1/get_month_spents": {
            "post": {
                "tags": [
                    "Invoices"
                ],
                "description": "Récupérer les factures du mois",
                "operationId": "3786fc0dce3f07ae576f651d6da03bb1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "month",
                                    "year"
                                ],
                                "properties": {
                                    "month": {
                                        "type": "integer"
                                    },
                                    "year": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Utilisateurs récupérés avec succès"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
...
    },
...
}